web/requirements.txt
asgiref==3.5.0
dj-rest-auth==2.1.11
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
